cube api into cdap-api
