+++ cdap-api/src/main/java/co/cask/cdap/api/dataset/lib/cube/CubeExploreQuery.java
@@ -14,33 +14,33 @@
  * the License.
  */
 
-package co.cask.cdap.metrics.store.cube;
+package co.cask.cdap.api.dataset.lib.cube;
 
+import co.cask.cdap.api.annotation.Beta;
 import com.google.common.base.Joiner;
 import com.google.common.base.Objects;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Maps;
 
-import java.util.Map;
+import java.util.List;
 
 /**
- * Query that specifies parameters to delete entries from cube.
+ * Defines a query to perform Exploration and Search on {@link Cube} data.
+ * Given a list of {@link TagValue} this explore query can be used
+ * to find next set of tags available or the measureNames belonging to this tag list.
  */
-public class CubeDeleteQuery {
-
+@Beta
+public class CubeExploreQuery {
   private final long startTs;
   private final long endTs;
   private final int resolution;
-  private final String measureName;
-  private final Map<String, String> sliceByTagValues;
+  private final int limit;
+  private final List<TagValue> tagValues;
 
-  public CubeDeleteQuery(long startTs, long endTs, int resolution, String measureName,
-                         Map<String, String> sliceByTagValues) {
+  public CubeExploreQuery(long startTs, long endTs, int resolution, int limit, List<TagValue> tagValues) {
     this.startTs = startTs;
     this.endTs = endTs;
     this.resolution = resolution;
-    this.measureName = measureName;
-    this.sliceByTagValues = Maps.newHashMap(sliceByTagValues);
+    this.limit = limit;
+    this.tagValues = tagValues;
   }
 
   public long getStartTs() {
@@ -55,12 +55,12 @@ public class CubeDeleteQuery {
     return resolution;
   }
 
-  public String getMeasureName() {
-    return measureName;
+  public int getLimit() {
+    return limit;
   }
 
-  public Map<String, String> getSliceByTags() {
-    return sliceByTagValues;
+  public List<TagValue> getTagValues() {
+    return tagValues;
   }
 
   @Override
@@ -68,8 +68,7 @@ public class CubeDeleteQuery {
     return Objects.toStringHelper(this)
       .add("startTs", startTs)
       .add("endTs", endTs)
-      .add("measureName", measureName)
-      .add("sliceByTags", Joiner.on(",").withKeyValueSeparator(":").useForNull("null").join(sliceByTagValues))
-      .toString();
+      .add("resolution", resolution)
+      .add("tagValues", Joiner.on(",").join(tagValues)).toString();
   }
 }
