+++ cdap-api/src/main/java/co/cask/cdap/api/dataset/lib/cube/TimeValue.java
@@ -14,14 +14,15 @@
  * the License.
  */
 
-package co.cask.cdap.api.metrics;
+package co.cask.cdap.api.dataset.lib.cube;
 
+import co.cask.cdap.api.annotation.Beta;
 import com.google.common.base.Objects;
 
 /**
  * Represents a value of the measure at specific timestamp.
  */
-//todo: move in higher-level package? It is used everywhere
+@Beta
 public final class TimeValue implements Comparable<TimeValue> {
   private final long timestamp;
   private final long value;
