+++ cdap-watchdog-api/src/main/java/co/cask/cdap/api/metrics/MetricDataQuery.java
@@ -16,10 +16,10 @@
 
 package co.cask.cdap.api.metrics;
 
+import co.cask.cdap.api.dataset.lib.cube.Interpolator;
 import com.google.common.base.Joiner;
 import com.google.common.base.Objects;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 
 import java.util.List;
