+++ cdap-watchdog-api/src/main/java/co/cask/cdap/api/metrics/MetricStore.java
@@ -16,6 +16,8 @@
 
 package co.cask.cdap.api.metrics;
 
+import co.cask.cdap.api.dataset.lib.cube.TagValue;
+
 import java.util.Collection;
 
 /**
