+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/cube/Aggregation.java
@@ -16,29 +16,34 @@
 
 package co.cask.cdap.metrics.store.cube;
 
+import co.cask.cdap.api.dataset.lib.cube.CubeFact;
+
 import java.util.List;
 
 /**
  * Defines an aggregation in {@link co.cask.cdap.metrics.store.cube.DefaultCube}.
  * <p/>
- * Aggregation usually defines a list of tags, or dimensions that are extracted from the {@link CubeFact} when it is
- * added to the {@link Cube}. Configuring a {@link Cube} to compute an aggregation allows later querying for data based
- * on those tags the data is aggregated for. See also {@link CubeQuery}.
+ * Aggregation usually defines a list of tags, or dimensions that are extracted from the
+ * {@link co.cask.cdap.api.dataset.lib.cube.CubeFact} when it is
+ * added to the {@link co.cask.cdap.api.dataset.lib.cube.Cube}. Configuring a
+ * {@link co.cask.cdap.api.dataset.lib.cube.Cube} to compute an aggregation allows later querying for data based
+ * on those tags the data is aggregated for. See also {@link co.cask.cdap.api.dataset.lib.cube.CubeQuery}.
  */
 public interface Aggregation {
   /**
    * Defines the tags to do the aggregation for.
    * <p/>
-   * Depending on the implementation of a {@link Cube}, the order of the tag names can be used as is to form a row key.
+   * Depending on the implementation of a {@link co.cask.cdap.api.dataset.lib.cube.Cube},
+   * the order of the tag names can be used as is to form a row key.
    * That in turn may affect the performance of the querying. Usually you want to put most frequently defined tags
-   * (tags with values to slice by in {@link CubeQuery} to be in front.
+   * (tags with values to slice by in {@link co.cask.cdap.api.dataset.lib.cube.CubeQuery} to be in front.
    *
    * @return list of tags (dimensions) to aggregate by
    */
   List<String> getTagNames();
 
   /**
-   * Filters out {@link CubeFact}s which should not be added into this aggregation.
+   * Filters out {@link co.cask.cdap.api.dataset.lib.cube.CubeFact}s which should not be added into this aggregation.
    * @param fact fact to test
    * @return true if fact should be added to the aggregation, false otherwise.
    */
