+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/timeseries/Fact.java
@@ -16,8 +16,9 @@
 
 package co.cask.cdap.metrics.store.timeseries;
 
-import co.cask.cdap.api.metrics.TagValue;
-import co.cask.cdap.api.metrics.TimeValue;
+import co.cask.cdap.api.dataset.lib.cube.MeasureType;
+import co.cask.cdap.api.dataset.lib.cube.TagValue;
+import co.cask.cdap.api.dataset.lib.cube.TimeValue;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
