+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/timeseries/FactCodec.java
@@ -17,7 +17,7 @@
 package co.cask.cdap.metrics.store.timeseries;
 
 import co.cask.cdap.api.common.Bytes;
-import co.cask.cdap.api.metrics.TagValue;
+import co.cask.cdap.api.dataset.lib.cube.TagValue;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import org.slf4j.Logger;
