+++ cdap-watchdog/src/test/java/co/cask/cdap/metrics/store/timeseries/FactCodecTest.java
@@ -16,7 +16,7 @@
 
 package co.cask.cdap.metrics.store.timeseries;
 
-import co.cask.cdap.api.metrics.TagValue;
+import co.cask.cdap.api.dataset.lib.cube.TagValue;
 import co.cask.cdap.common.utils.ImmutablePair;
 import co.cask.cdap.data2.dataset2.lib.table.FuzzyRowFilter;
 import co.cask.cdap.data2.dataset2.lib.table.MetricsTable;
